scx_bpfland: improve vruntime / time slice evaluation
Always assign the maximum time slice if there are idle CPUs in the
system. Otherwise, double the task's unused time slice to reward tasks
that use less CPU time and at the same time refill the time slice of the
tasks every time they're dispatched.

Moreover, make sure that the minimum vruntime of a task is periodically
refreshed, also when the task is directly dispatched.

Signed-off-by: Andrea Righi <righi.andrea@gmail.com>